// Samples/typescript/wizardpin/SigCaptX-WizardPINPad-PadDefs.d.ts
export declare class PadControl {
    buttonSize: number;
    buttonWidth: number;
    font: string;
    height: number;
    width: number;
    range: string;
    textSize: number;
    titleBold: boolean;
    type: string;
    xButtonCancel: any;
    xButtonNext: any;
    yButtonCancel: any;
    yButtonNext: any;
    yLSText: number;
    yText: number;
    constructor(width: any, height: any);
}
export declare class Screen_Display1 {
    cancelButton: any;
    enterBelow: any;
    keyWidth: number;
    nextButton: any;
    pin1: any;
    pin2: any;
    pin3: any;
    pin4: any;
    pin5: any;
    pin6: any;
    pin7: any;
    pin8: any;
    pin9: any;
    yInputEcho: number;
    xButtonCol1: number;
    xButtonCol2: number;
    xButtonCol3: number;
    yButtonRow1: number;
    yButtonRow2: number;
    yButtonRow3: number;
    constructor(pad: any);
}